slimming down keyboard stuff, backlight breathing implemented
