+++ tmk_core/common/matrix.h
@@ -76,6 +76,13 @@ void matrix_power_down(void);
 /* executes code for Quantum */
 void matrix_init_quantum(void);
 void matrix_scan_quantum(void);
+
+void matrix_init_kb(void);
+void matrix_scan_kb(void);
+
+void matrix_init_user(void);
+void matrix_scan_user(void);
+
 #ifdef __cplusplus
 }
 #endif
